+++ pkg/resource/events-logger-config/alloy-events-config.go
@@ -77,6 +77,8 @@ func generateAlloyConfig(cluster *capi.Cluster, includeNamespaces []string, excl
 		IsWorkloadCluster  bool
 		TracingEnabled     bool
 		TracingEndpoint    string
+		TracingUsernameKey string
+		TracingPasswordKey string
 		Tenants            []string
 	}{
 		ClusterID:          cluster.GetName(),
@@ -94,6 +96,8 @@ func generateAlloyConfig(cluster *capi.Cluster, includeNamespaces []string, excl
 		IsWorkloadCluster:  common.IsWorkloadCluster(installationName, cluster.GetName()),
 		TracingEnabled:     tracingEnabled,
 		TracingEndpoint:    endpoint,
+		TracingUsernameKey: common.TracingUsername,
+		TracingPasswordKey: common.TracingPassword,
 		Tenants:            tenants,
 	}