+++ pkg/resource/logging-secret/alloy-logging-secret.go
@@ -12,11 +12,11 @@ import (
 	"github.com/Masterminds/sprig/v3"
 
 	"github.com/giantswarm/logging-operator/pkg/common"
-	loggingcredentials "github.com/giantswarm/logging-operator/pkg/resource/logging-credentials"
+	credentials "github.com/giantswarm/logging-operator/pkg/resource/credentials"
 )
 
 var (
-	//go:embed alloy/logging-secret.yaml.template
+	//go:embed alloy/alloy-secret.yaml.template
 	alloySecret         string
 	alloySecretTemplate *template.Template
 )
@@ -25,10 +25,11 @@ func init() {
 	alloySecretTemplate = template.Must(template.New("logging-secret.yaml").Funcs(sprig.FuncMap()).Parse(alloySecret))
 }
 
-func GenerateAlloyLoggingSecret(cluster *capi.Cluster, credentialsSecret *v1.Secret, lokiURL string) (map[string][]byte, error) {
+func GenerateAlloyLoggingSecret(cluster *capi.Cluster, credentialsSecret *v1.Secret, lokiURL string, tracingEnabled bool, tracingCredentialsSecret *v1.Secret) (map[string][]byte, error) {
 	clusterName := cluster.GetName()
+	var values bytes.Buffer
 
-	writePassword, err := loggingcredentials.GetPassword(cluster, credentialsSecret, clusterName)
+	writePassword, err := credentials.GetPassword(cluster, credentialsSecret, clusterName)
 	if err != nil {
 		return nil, err
 	}
@@ -45,7 +46,16 @@ func GenerateAlloyLoggingSecret(cluster *capi.Cluster, credentialsSecret *v1.Sec
 		},
 	}
 
-	var values bytes.Buffer
+	if tracingEnabled {
+		tracingPassword, err := credentials.GetPassword(cluster, tracingCredentialsSecret, clusterName)
+		if err != nil {
+			return nil, err
+		}
+
+		templateData.ExtraSecretEnv[common.TracingUsername] = clusterName
+		templateData.ExtraSecretEnv[common.TracingPassword] = tracingPassword
+	}
+
 	err = alloySecretTemplate.Execute(&values, templateData)
 	if err != nil {
 		return nil, err